// pecus.Frontend/src/connectors/api/pecus/models/TaskType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaskType = 'Bug' | 'Feature' | 'Documentation' | 'Review' | 'Testing' | 'Refactoring' | 'Research' | 'Meeting' | 'BusinessNegotiation' | 'RequirementsConfirmation' | 'Other';